+++ Before.Java.Interop.sln.targets
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildThisFileDirectory)build-tools\scripts\PrepareWindows.targets" Condition=" '$(OS)' == 'Windows_NT' " />
+  <Import Project="$(MSBuildThisFileDirectory)build-tools\scripts\RunNUnitTests.targets" />
 </Project>
