--- tools/generator/Tests/expected.ji/InterfaceMethodsConflict/Java.Lang.Object.cs
@@ -1,13 +1,0 @@
-using System;
-using System.Collections.Generic;
-using Android.Runtime;
-using Java.Interop;
-
-namespace Java.Lang {
-
-	// Metadata.xml XPath class reference: path="/api/package[@name='java.lang']/class[@name='Object']"
-	[global::Android.Runtime.Register ("java/lang/Object", DoNotGenerateAcw=true)]
-	public partial class Object  {
-
-	}
-}
